Python/scripts/requirements.txt
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.5
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mne==1.11.0
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyxdf==1.17.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.6.0
